Move header.line to line.py, rename HeaderLine
Avoiding name conflicts turns out to be useful for cross-referencing
class in the docstrings (when ran through sphinx). Since the HeaderLine
really does implement a line interface, move it to the approperiate
file.
