+++ python/segyio/trace.py
@@ -6,7 +6,7 @@ except ImportError: pass
 import numpy as np
 
 from ._raw_trace import RawTrace
-from .line import Line as LineBase
+from .line import HeaderLine
 from .field import Field
 from .utils import castarray
 
@@ -327,7 +327,14 @@ class Header(Sequence):
 
     @property
     def iline(self):
-        return Line(self, self.segy.iline, 'inline')
+        """
+        Headers, accessed by inline
+
+        Returns
+        -------
+        line : HeaderLine
+        """
+        return HeaderLine(self, self.segy.iline, 'inline')
 
     @iline.setter
     def iline(self, value):
@@ -345,7 +352,14 @@ class Header(Sequence):
 
     @property
     def xline(self):
-        return Line(self, self.segy.xline, 'crossline')
+        """
+        Headers, accessed by crossline
+
+        Returns
+        -------
+        line : HeaderLine
+        """
+        return HeaderLine(self, self.segy.xline, 'crossline')
 
     @xline.setter
     def xline(self, value):
@@ -361,153 +375,3 @@ class Header(Sequence):
 
         for i, src in zip(self.segy.xlines, value):
             self.xline[i] = src
-
-class Line(LineBase):
-    # a lot of implementation details are shared between reading data traces
-    # line-by-line and trace headers line-by-line, so (ab)use inheritance for
-    # __len__, keys() etc., however, the __getitem__ is  way different and is re-implemented
-
-    def __init__(self, header, base, direction):
-        super(Line, self).__init__(header.segy,
-                                   base.lines,
-                                   base.length,
-                                   base.stride,
-                                   sorted(base.offsets.keys()),
-                                   'header.' + direction,
-                                  )
-        self.header = header
-
-    def __getitem__(self, index):
-        """line[i] or line[i, o]
-
-        The line `i`, or the line `i` at a specific offset `o`. ``line[i]``
-        returns an iterable of `Field` objects, and changes to these *will* be
-        reflected on disk.
-
-        The `i` and `o` are *keys*, and should correspond to the line- and
-        offset labels in your file, and in the `ilines`, `xlines`, and
-        `offsets` attributes.
-
-        Slices can contain lines and offsets not in the file, and like with
-        list slicing, these are handled gracefully and ignored.
-
-        When `i` or `o` is a slice, a generator of iterables of headers are
-        returned.
-
-        When both `i` and `o` are slices, one generator is returned for the
-        product `i` and `o`, and the lines are yielded offsets-first, roughly
-        equivalent to the double for loop::
-
-            >>> for line in lines:
-            ...     for off in offsets:
-            ...         yield line[line, off]
-            ...
-
-        Parameters
-        ----------
-
-        i : int or slice
-        o : int or slice
-
-        Returns
-        -------
-
-        line : iterable of Field or generator of iterator of Field
-
-        Raises
-        ------
-
-        KeyError
-            If `i` or `o` don't exist
-
-        Notes
-        -----
-
-        .. versionadded:: 1.1
-
-        """
-        offset = self.default_offset
-        try: index, offset = index
-        except TypeError: pass
-
-        try:
-            start = self.heads[index] + self.offsets[offset]
-        except TypeError:
-            # index is either unhashable (because it's a slice), or offset is a
-            # slice.
-            pass
-
-        else:
-            step = self.stride * len(self.offsets)
-            return self.header[start::step]
-
-        def gen():
-            irange, orange = self.ranges(index, offset)
-            for line in irange:
-                for off in orange:
-                    yield self[line, off]
-
-        return gen()
-
-    def __setitem__(self, index, val):
-        """line[i] = val or line[i, o] = val
-
-        Follows the same rules for indexing and slicing as ``line[i]``. If `i`
-        is an int, and `val` is a dict or Field, that value is replicated and
-        assigned to every trace header in the line, otherwise it's treated as
-        an iterable, and each trace in the line is assigned the ``next()``
-        yielded value.
-
-        If `i` or `o` is a slice, `val` must be an iterable.
-
-        In either case, if the `val` iterable is exhausted before the line(s),
-        assignment stops with whatever is written so far.
-
-        Parameters
-        ----------
-
-        i : int or slice
-        offset : int or slice
-        val : field or dict_like or iterable of field or iterable of dict_like
-
-        Raises
-        ------
-
-        KeyError
-            If `i` or `o` don't exist
-
-        Notes
-        -----
-
-        .. versionadded:: 1.1
-
-        Behaves like [] for lists.
-
-        """
-        offset = self.default_offset
-        try: index, offset = index
-        except TypeError: pass
-
-        try: start = self.heads[index] + self.offsets[offset]
-        except TypeError: pass
-
-        else:
-            try:
-                if hasattr(val, 'keys'):
-                    val = itertools.repeat(val)
-            except TypeError:
-                # already an iterable
-                pass
-
-            step = self.stride * len(self.offsets)
-            self.header[start::step] = val
-            return
-
-        irange, orange = self.ranges(index, offset)
-        val = iter(val)
-        for line in irange:
-            for off in orange:
-                try:
-                    self[line, off] = next(val)
-                except StopIteration:
-                    return
